Test Greater Equal: Adds tests for result dtype and moves tests to separate file
